Not fun part: Pseudo-random generator configuration
That would be the seed of the pseudo-random generator. As I do not want
to commit the seed I personally use, I leave that to be filled
up by anyone that wants to use this tool. Unconvenient, I know, because
the numbers will not be publicly available if you lose them...