Fix warnings when built with `-Wconversion`
Use a lot of `size_t` or `unsigned` when appropriate.

Also, add casts when needed. Not sure this is actually better, though,
but when the output of `atoi()` must be stored in an `unsigned int`,
there is no other possibility. Well, `atoi()` is used only to convert
the year input... Actual issues may appear if you still use this tool
by year 2147483648. Or later, maybe `int` will not be on 32-bit any
more by this time...

Oh, build with `-Wconversion` now, of course.

Fix #3